Central_Telemetry_Handler/requirements.txt
Flask
flask-httpauth
pymongo[srv]
PyQt5
requests
pyyaml
